move no content available component to it's own folder
