+++ src/DetailsView/components/no-content-available/no-content-available.tsx
